# swagger.json
{
"swagger": "2.0",
"info": {
"description": "go-gin swagger 文档测试.",
"title": "还只是个测试",
"contact": {
"name": "zw",
"email": "wei.zhou@shunwang.com"
},
"license": {}
},
"paths": {
"/article": {
"get": {
"description": "文章列表",
"consumes": [
"application/json"
],
"produces": [
"application/json"
],
"tags": [
"article"
],
"summary": "文章列表",
"parameters": [
{
"type": "string",
"description": "page",
"name": "page",
"in": "query"
},
{
"type": "string",
"description": "limit",
"name": "limit",
"in": "query"
}
],
"responses": {
"200": {
"description": "文章列表",
"schema": {
"type": "object",
"$ref": "#/definitions/Articles"
}
}
}
},
"post": {
"description": "Post",
"consumes": [
"application/json"
],
"produces": [
"application/json"
],
"tags": [
"article"
],
"summary": "Post",
"parameters": [
{
"description": "pets fields",
"name": "pets",
"in": "body",
"required": true,
"schema": {
"type": "object",
"$ref": "#definitions/Article"
}
}
],
"responses": {
"200": {
"description": "success",
"schema": {
"type": "object",
"$ref": "#/definitions/Article"
}
}
}
}
},
"/thing": {
"get": {
"description": "单个",
"consumes": [
"application/json"
],
"produces": [
"application/json"
],
"tags": [
"users"
],
"summary": "单个",
"parameters": [
{
"type": "string",
"description": "_id of doc",
"name": "_id",
"in": "query"
}
],
"responses": {
"200": {
"description": "doc",
"schema": {
"type": "object",
"$ref": "#/definitions/Pets"
}
}
}
}
},
"/things": {
"get": {
"description": "列表",
"consumes": [
"text/html"
],
"produces": [
"application/json"
],
"tags": [
"users"
],
"summary": "列表",
"parameters": [
{
"type": "string",
"description": "page",
"name": "page",
"in": "query"
}
],
"responses": {
"200": {
"description": "petslist",
"schema": {
"type": "object",
"$ref": "#/definitions/Things"
}
}
}
},
"put": {
"description": "Put",
"consumes": [
"application/json"
],
"produces": [
"application/json"
],
"tags": [
"users"
],
"summary": "Put",
"parameters": [
{
"description": "pets fields",
"name": "pets",
"in": "body",
"required": true,
"schema": {
"type": "object",
"$ref": "#definitions/Thing"
}
}
],
"responses": {
"200": {
"description": "success",
"schema": {
"type": "object",
"$ref": "#/definitions/Thing"
}
}
}
},
"post": {
"description": "新增",
"consumes": [
"application/json"
],
"produces": [
"application/json"
],
"tags": [
"users"
],
"summary": "新增",
"parameters": [
{
"description": "pets fields",
"name": "pets",
"in": "body",
"required": true,
"schema": {
"type": "object",
"$ref": "#definitions/Thing"
}
}
],
"responses": {
"200": {
"description": "success",
"schema": {
"type": "object",
"$ref": "#/definitions/Thing"
}
}
}
},
"delete": {
"description": "Delete",
"consumes": [
"text/html"
],
"produces": [
"application/json"
],
"tags": [
"users"
],
"summary": "Delete",
"parameters": [
{
"type": "string",
"description": "_id of doc",
"name": "_id",
"in": "query"
}
],
"responses": {
"200": {
"description": "doc",
"schema": {
"type": "object",
"$ref": "#/definitions/Pets"
}
}
}
}
}
},
"definitions": {
"Article": {
"type": "object",
"properties": {
"content": {
"type": "string"
},
"created_at": {
"type": "number"
},
"id": {
"type": "number"
},
"title": {
"type": "string"
},
"updated_at": {
"type": "number"
}
}
},
"Articles": {
"type": "object",
"properties": {
"data": {
"type": "array"
}
}
},
"Error": {
"type": "object",
"properties": {
"code": {
"type": "int"
},
"message": {
"type": "string"
}
}
},
"Pets": {
"type": "object",
"properties": {
"id": {
"type": "string"
},
"tag": {
"type": "array"
}
}
},
"Tag": {
"type": "object",
"properties": {
"id": {
"type": "string"
},
"name": {
"type": "string"
}
}
},
"Thing": {
"type": "object",
"properties": {
"id": {
"type": "string"
},
"name": {
"type": "string"
},
"value": {
"type": "string"
}
}
},
"Things": {
"type": "object",
"properties": {
"data": {
"type": "array"
}
}
}
}
}